계속 에러뜨는 이유를 찾았다
