+++ documents/태욱이가 보내준 링크.md
@@ -66,3 +66,8 @@ export DATABASE_URL=(디비 링크)
 
 
 
+이거 sequelize 버전이 3.23.4다....
+
+현재(2020-08-11) 6.3.4임....
+
+이정도면 디프리케이티드를 넘어서 기록에도 안남을정도 아닐까
