--- postgres-express-react-node-tutorial(gone_wrong)/.sequelizerc
@@ -1,8 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  "config": path.resolve('./server/config', 'config.json'),
-  "models-path": path.resolve('./server/models'),
-  "seeders-path": path.resolve('./server/seeders'),
-  "migrations-path": path.resolve('./server/migrations')
-};
