--- postgres-express-react-node-tutorial(gone_wrong)/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "postgres-express-react-node-tutorial",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start:dev": "nodemon ./bin/www",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "morgan": "^1.10.0",
-    "pg": "^8.3.0",
-    "pg-hstore": "^2.3.3",
-    "sequelize": "^6.3.3"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.4"
-  }
-}
