--- postgres-express-react-node-tutorial(gone_wrong)/server/config/config.json
@@ -1,19 +1,0 @@
-{
-  "development": {
-    "use_env_variable": "DATABASE_URL"
-  },
-  "test": {
-    "username": "ssuw2",
-    "password": "1234",
-    "database": "todos-test",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  },
-  "production": {
-    "username": "ssuw2",
-    "password": "1234",
-    "database": "todos-pro",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  }
-}
