--- postgres-express-react-node-tutorial(gone_wrong)/server/controllers/index.js
@@ -1,5 +1,0 @@
-const todos = require('./todos');
-
-module.exports = {
-  todos,
-};
