--- postgres-express-react-node-tutorial(gone_wrong)/server/controllers/todos.js
@@ -1,12 +1,0 @@
-const Todo = require('../models').Todo;
-
-module.exports = {
-  create(req, res) {
-    return Todo
-      .create({
-        title: req.body.title,
-      })
-      .then(todo => res.status(201).send(todo))
-      .catch(error => res.status(400).send(error));
-  },
-};
