--- postgres-express-react-node-tutorial(gone_wrong)/server/models/todo.js
@@ -1,17 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-  const Todo = sequelize.define('Todo', {
-    title: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  });
-
-  Todo.associate = (models) => {
-    Todo.hasMany(models.TodoItem, {
-      foreignKey: 'todoId',
-      as: 'todoItems',
-    });
-  };
-
-  return Todo;
-};
