--- postgres-express-react-node-tutorial(gone_wrong)/server/models/todoitem.js
@@ -1,21 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-  const TodoItem = sequelize.define('TodoItem', {
-    content: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    complete: {
-      type: DataTypes.BOOLEAN,
-      defaultValue: false,
-    },
-  });
-
-  TodoItem.associate = (models) => {
-    TodoItem.belongsTo(models.Todo, {
-      foreignKey: 'todoId',
-      onDelete: 'CASCADE',
-    });
-  };
-
-  return TodoItem;
-};
