--- postgres-express-react-node-tutorial(gone_wrong)/server/routes/index.js
@@ -1,9 +1,0 @@
-const todosController = require('../controllers').todos;
-
-module.exports = (app) => {
-  app.get('/api', (req, res) => res.status(200).send({
-    message: 'Welcome to the Todos API!',
-  }));
-
-  app.post('/api/todos', todosController.create);
-};
