--- postgres-express-react-node-tutorial/server/config/index.js
@@ -1,31 +1,0 @@
-require('dotenv').config();
-const env = process.env;
-
-const development = {
-  username: env.MYSQL_USERNAME,
-  password: env.MYSQL_PASSWORD,
-  database: env.MYSQL_DATABASE,
-  host: env.MYSQL_HOST,
-  dialect: env.MYSQL_DIALECT,
-  port: env.MYSQL_PORT
-};
-
-const production = {
-  username: env.MYSQL_USERNAME,
-  password: env.MYSQL_PASSWORD,
-  database: env.MYSQL_DATABASE,
-  host: env.MYSQL_HOST,
-  dialect: env.MYSQL_DIALECT,
-  port: env.MYSQL_PORT
-};
-
-const test = {
-  username: env.MYSQL_USERNAME,
-  password: env.MYSQL_PASSWORD,
-  database: env.MYSQL_DATABASE_TEST,
-  host: env.MYSQL_HOST,
-  dialect: env.MYSQL_DIALECT,
-  port: env.MYSQL_PORT
-};
-
-module.exports = { development, production, test };
