+++ postgres-express-react-node-tutorial/server/controllers/index.js
@@ -1,5 +1,7 @@
 const todos = require('./todos');
+// const todoItems = require('./todoitems');
 
 module.exports = {
   todos,
+  todoItems,
 };
