+++ postgres-express-react-node-tutorial/server/models/todo.js
@@ -4,14 +4,15 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.STRING,
       allowNull: false,
     },
+  }, {
+    classMethods: {
+      associate: (models) => {
+        Todo.hasMany(models.TodoItem, {
+          foreignKey: 'todoId',
+          as: 'todoItems',
+        });
+      },
+    },
   });
-
-  Todo.associate = (models) => {
-    Todo.hasMany(models.TodoItem, {
-      foreignKey: 'todoId',
-      as: 'todoItems',
-    });
-  };
-
   return Todo;
 };
