+++ postgres-express-react-node-tutorial/server/models/todoitem.js
@@ -8,14 +8,15 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.BOOLEAN,
       defaultValue: false,
     },
+  }, {
+    classMethods: {
+      associate: (models) => {
+        TodoItem.belongsTo(models.Todo, {
+          foreignKey: 'todoId',
+          onDelete: 'CASCADE',
+        });
+      },
+    },
   });
-
-  TodoItem.associate = (models) => {
-    TodoItem.belongsTo(models.Todo, {
-      foreignKey: 'todoId',
-      onDelete: 'CASCADE',
-    });
-  };
-
   return TodoItem;
 };
