+++ postgres-express-react-node-tutorial/server/routes/index.js
@@ -1,4 +1,5 @@
 const todosController = require('../controllers').todos;
+const todoItemsController = require('../controllers').todoItems;
 
 module.exports = (app) => {
   app.get('/api', (req, res) => res.status(200).send({
